% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactTool.R
\name{relativeProductImpact}
\alias{relativeProductImpact}
\title{Relative Product Impact Rank}
\usage{
relativeProductImpact(productCodes, dateStart, dateEnd)
}
\arguments{
\item{productCodes}{vector of product_codes}

\item{dateStart}{date from which kpi's start being measured format yyyy-mm-dd}

\item{dateEnd}{date to which kpi's stop being measured format yyyy-mm-dd}
}
\value{
data frame containing product_codes and their relative impact
}
\description{
The function impact() takes a vector of product codes and returns each product's relative impact -- a measure that reflects a product's conversion, weighted by its daily TTV.
}
\examples{
relativeProductImpact(c(404805, 544419, 20072), '2016-08-09', '2016-08-23')

}
\author{
Lee Stirling, Kathryn Harris 2016-08-24
}

